run-travel-ai-agent/requirements.txt
fastapi
uvicorn
requests
google-generativeai
python-dotenv
httpx
asyncio
google-cloud-logging
adk